Smaller changes - Added change column type to table viewer's context menu - Disabled the transparency of the PNG output in graph viewer - Resolved a problem with to_cmd() - Resolved a size_t issue - Added the architecture to getversioninfo()

+++ kernel/core/procedure/flowctrl.hpp
@@ -191,7 +191,7 @@ class FlowCtrl
         void eval();
         void reset();
 
-        size_t getCurrentLineNumber() const;
+        int getCurrentLineNumber() const;
         std::string getCurrentCommand() const;
 
         static bool isFlowCtrlStatement(const std::string& sCmd);
